pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pi-drone-nav"
version = "0.1.0"
description = "Autonomous drone navigation for Raspberry Pi with Betaflight"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
full = [
    "flask>=2.0.0",
    "flask-cors>=3.0.0",
    "pymavlink>=2.4.0",
    "pyyaml>=6.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
pi-drone-nav = "src.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/pi-drone-nav"
Repository = "https://github.com/yourusername/pi-drone-nav"
Issues = "https://github.com/yourusername/pi-drone-nav/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
